requirements.txt
greenlet==3.1.1
mysql-connector-python==9.2.0
playwright==1.49.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyee==12.0.0
python-decouple==3.8
typing_extensions==4.12.2
